Change CreatedAfter2 type from long? to int? in TimeFilterRequest.

+++ Apps.Hubspot/Models/Requests/TimeFilterRequest.cs
@@ -19,7 +19,7 @@ public class TimeFilterRequest
     [Display("Created after milliseconds")]
     [JsonProperty("createdAt__gt")]
     //[JsonConverter(typeof(DateTimeToUnixEpoch))]
-    public long? CreatedAfter2 { get; set; }
+    public int? CreatedAfter2 { get; set; }
 
     [Display("Created before")]
     [JsonProperty("createdAt__lt")]
